fix: Set env variable for hostname argument of tideways daemon
The new daemon version only adds the container id of the running daemon
to the hostname. This results in something like
tideways-daemon#4bae42711daa, which is not working for us.
After deployments on live and staging, a new tideways-daemon found its
way into the list of servers in tideways with a new hostname (see screenshot in mentioned
ticket) which resulted in a total of 32 servers for the related project.