Remove EXPOSE for port 9145 in Dockerfile
Remove EXPOSE directive for port 9145 in Dockerfile.

If you want to run this exporter container multiple times, e.g. in an multi docker compose setup or with multiple nginx instances as proxy and webserver, this port will always be exposed.
The prometheus takes the first located port in service discovery, therefore the target is down and no metrics are collected.