fix: Don't configure php error_reporting through env variable
PHP won't parse E_* constants when they're not set directly in the ini.
Because of this, the previous value would evaluate to 0 during runtime,
effectively disabling all error reporting.

This change reverts the error_reporting back to the old default. If needed,
the error_reporting setting can be override by a new ini file placed in
the conf.d file by the consumer Dockerfile